// frontend/src/api/generated/models/getBudgetOverviewFinanceBudgetsGetParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable ticket tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */

export type GetBudgetOverviewFinanceBudgetsGetParams = {
/**
 * The period's first day
 */
start: string;
/**
 * The period's last day
 */
end: string;
};